Move EventStreamId into event_store module
It is not specific to the Kurrent adapter.
